requirements.txt
dash_core_components==1.0.0
pkuseg==0.0.22
plotly==4.8.1
wordcloud==1.7.0
pandas==0.25.3
dash_html_components==1.0.0
imageio==2.8.0
dash_table==4.0.2
dash==1.0.2
pytz==2020.1
numpy==1.18.5
mysql-connector-python==8.0.19
python-dateutil==2.6.1
